backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=src/requirements.txt src/requirements/requirements.in
#
asgiref==3.7.2
    # via django
django==4.1.10
    # via
    #   -r src/requirements/requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.2.0
    # via -r src/requirements/requirements.in
djangorestframework==3.14.0
    # via -r src/requirements/requirements.in
pillow==10.0.0
    # via -r src/requirements/requirements.in
pytz==2023.3
    # via djangorestframework
sqlparse==0.4.4
    # via django
tzdata==2023.3
    # via django
